removing duplicate paths
